requirements.txt
fastapi[all]
uvicorn
langchain
langchain-community
langchain-google-genai # <--- New: For Gemini integration
sqlalchemy
python-dotenv
# Add your specific database connector:
# psycopg2-binary
# mysql-connector-python
# pymssql
pydantic